pyproject.toml
[project]
name = "actions-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.9.1",
    "dbt-duckdb>=1.9.1",
    "dlt[motherduck]>=1.5.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
